Fix ReSharper lint issues
- Remove unused 'count' variables in RepositorySyncJob
- Remove unused 'gpgPending' and 'neverSynced' variables in Mirrors/Index view
- Remove redundant type cast in AptVersionComparisonService
- Remove unused 'using Microsoft.Extensions.Logging' in RepositoryDependencyCheckJob
- Use discard parameter '_' for unused cancellationToken
- Add ReSharper suppression comment for MissingDeps (populated via JSON deserialization)
- Move script from @section to inline (section not defined in layout)

+++ src/Aiursoft.Apkg/Services/AptVersionComparisonService.cs
@@ -93,7 +93,7 @@ public partial class AptVersionComparisonService
                 {
                     int ac = i < part1.Length ? (int)char.GetNumericValue(part1[i]) : -1;
                     int bc = j < part2.Length ? (int)char.GetNumericValue(part2[j]) : -1;
-                    firstDiff = (int)(ac - bc);
+                    firstDiff = ac - bc;
                 }
 
                 if (i < part1.Length && char.IsDigit(part1[i])) i++;
